refactor(visualization): turn the file loaders into the features/loadFile slice
Move loadFile.service and loadInitialFile.service out of services/ into
features/loadFile (the services/ folder is now gone), behind a facade.

- Full ngrx stores refactor: only stores/ touch @ngrx/store. LoadFileStore
  wraps loadFile's few selects/dispatches; LoadInitialFileStore holds the
  ~53 dispatches that hydrate the whole store from a saved file/URL/IndexedDB
  (the apply*/map* settings logic moved in verbatim). The services keep the
  orchestration (flow, error dialogs, URL reading via UrlExtractor) and call
  store methods; loadInitialFile reads URL params and passes the values to
  LoadInitialFileStore.setMetricsFromUrlValues/setRenderState.
- facade.ts exposes LoadFileService, LoadInitialFileService, sampleFile1/2.
  Repoint navBar, globalSettings, codeCharta.component and the
  updateQueryParameters effect to the facade.

Behavior unchanged. tsc, full Jest suite (378 suites), dep-cruiser (0 errors)
and `ng build` (exit 0) pass.
