+++ visualization/app/codeCharta/features/globalSettings/components/globalConfigurationDialog/confirmResetMapDialog/confirmResetMapDialog.spec.ts
@@ -6,8 +6,8 @@ import { render, screen } from "@testing-library/angular"
 import userEvent from "@testing-library/user-event"
 import "fake-indexeddb/auto"
 import { getNameDataPair } from "../../../../../util/fileParser"
-import { LoadFileService } from "../../../../../services/loadFile/loadFile.service"
-import { LoadInitialFileService, sampleFile1, sampleFile2 } from "../../../../../services/loadInitialFile/loadInitialFile.service"
+import { LoadFileService } from "../../../../../features/loadFile/facade"
+import { LoadInitialFileService, sampleFile1, sampleFile2 } from "../../../../../features/loadFile/facade"
 import { UrlExtractor } from "../../../../../util/urlExtractor"
 import * as resetChosenMetricsEffect from "../../../../../state/effects/resetChosenMetrics/resetChosenMetrics.effect"
 import { metricDataSelector } from "../../../../../state/selectors/accumulatedData/metricData/metricData.selector"
