+++ visualization/app/codeCharta/features/loadFile/services/loadInitialFile.service.spec.ts
@@ -4,34 +4,34 @@ import { State, StoreModule } from "@ngrx/store"
 import { MockStore, provideMockStore } from "@ngrx/store/testing"
 import { waitFor } from "@testing-library/angular"
 import stringify from "safe-stable-stringify"
-import { AppSettings, CcState, DynamicSettings, FileSettings } from "../../codeCharta.model"
-import { FileSelectionState } from "../../model/files/files"
-import { getCCFiles } from "../../model/files/files.helper"
-import { MetricQueryParemter } from "../../state/effects/updateQueryParameters/metricQueryParameter"
-import { metricDataSelector } from "../../state/selectors/accumulatedData/metricData/metricData.selector"
-import { setAmountOfTopLabels } from "../../state/store/appSettings/amountOfTopLabels/amountOfTopLabels.actions"
-import { defaultAppSettings } from "../../state/store/appSettings/appSettings.reducer"
-import { setAreaMetric } from "../../state/store/dynamicSettings/areaMetric/areaMetric.actions"
-import { setColorMetric } from "../../state/store/dynamicSettings/colorMetric/colorMetric.actions"
-import { defaultDynamicSettings } from "../../state/store/dynamicSettings/dynamicSettings.reducer"
-import { setEdgeMetric } from "../../state/store/dynamicSettings/edgeMetric/edgeMetric.actions"
-import { setHeightMetric } from "../../state/store/dynamicSettings/heightMetric/heightMetric.actions"
-import { defaultFileSettings } from "../../state/store/fileSettings/fileSettings.reducer"
-import { setDelta, setFiles } from "../../state/store/files/files.actions"
-import { appReducers, defaultState, setStateMiddleware } from "../../state/store/state.manager"
-import { EDGE_METRIC_DATA, FILE_STATES, METRIC_DATA, TEST_DELTA_MAP_A, TEST_DELTA_MAP_B } from "../../util/dataMocks"
-import { readCcState } from "../../util/indexedDB/indexedDBWriter"
-import { getLastAction } from "../../util/testUtils/store.utils"
-import { ErrorDialogService } from "../../features/shared/components/errorDialog/errorDialog.service"
-import { getNameDataPair } from "../../util/fileParser"
-import { LoadFileService } from "../loadFile/loadFile.service"
+import { AppSettings, CcState, DynamicSettings, FileSettings } from "../../../codeCharta.model"
+import { FileSelectionState } from "../../../model/files/files"
+import { getCCFiles } from "../../../model/files/files.helper"
+import { MetricQueryParemter } from "../../../state/effects/updateQueryParameters/metricQueryParameter"
+import { metricDataSelector } from "../../../state/selectors/accumulatedData/metricData/metricData.selector"
+import { setAmountOfTopLabels } from "../../../state/store/appSettings/amountOfTopLabels/amountOfTopLabels.actions"
+import { defaultAppSettings } from "../../../state/store/appSettings/appSettings.reducer"
+import { setAreaMetric } from "../../../state/store/dynamicSettings/areaMetric/areaMetric.actions"
+import { setColorMetric } from "../../../state/store/dynamicSettings/colorMetric/colorMetric.actions"
+import { defaultDynamicSettings } from "../../../state/store/dynamicSettings/dynamicSettings.reducer"
+import { setEdgeMetric } from "../../../state/store/dynamicSettings/edgeMetric/edgeMetric.actions"
+import { setHeightMetric } from "../../../state/store/dynamicSettings/heightMetric/heightMetric.actions"
+import { defaultFileSettings } from "../../../state/store/fileSettings/fileSettings.reducer"
+import { setDelta, setFiles } from "../../../state/store/files/files.actions"
+import { appReducers, defaultState, setStateMiddleware } from "../../../state/store/state.manager"
+import { EDGE_METRIC_DATA, FILE_STATES, METRIC_DATA, TEST_DELTA_MAP_A, TEST_DELTA_MAP_B } from "../../../util/dataMocks"
+import { readCcState } from "../../../util/indexedDB/indexedDBWriter"
+import { getLastAction } from "../../../util/testUtils/store.utils"
+import { ErrorDialogService } from "../../../features/shared/components/errorDialog/errorDialog.service"
+import { getNameDataPair } from "../../../util/fileParser"
+import { LoadFileService } from "./loadFile.service"
 import { LoadInitialFileService, sampleFile1, sampleFile2 } from "./loadInitialFile.service"
-import { UrlExtractor } from "../../util/urlExtractor"
-import { setCurrentFilesAreSampleFiles } from "../../state/store/appStatus/currentFilesAreSampleFiles/currentFilesAreSampleFiles.actions"
+import { UrlExtractor } from "../../../util/urlExtractor"
+import { setCurrentFilesAreSampleFiles } from "../../../state/store/appStatus/currentFilesAreSampleFiles/currentFilesAreSampleFiles.actions"
 
-jest.mock("../../util/urlExtractor")
-jest.mock("../../model/files/files.helper")
-jest.mock("../../util/indexedDB/indexedDBWriter")
+jest.mock("../../../util/urlExtractor")
+jest.mock("../../../model/files/files.helper")
+jest.mock("../../../util/indexedDB/indexedDBWriter")
 
 describe("LoadInitialFileService", () => {
     let store: MockStore
