+++ visualization/app/codeCharta/features/navBar/components/settingsButton/settingsButton.component.spec.ts
@@ -4,8 +4,8 @@ import { State } from "@ngrx/store"
 import { provideMockStore } from "@ngrx/store/testing"
 import { screen } from "@testing-library/angular"
 import userEvent from "@testing-library/user-event"
-import { LoadFileService } from "../../../../services/loadFile/loadFile.service"
-import { LoadInitialFileService } from "../../../../services/loadInitialFile/loadInitialFile.service"
+import { LoadFileService } from "../../../../features/loadFile/facade"
+import { LoadInitialFileService } from "../../../../features/loadFile/facade"
 import { defaultAppSettings } from "../../../../state/store/appSettings/appSettings.reducer"
 import { GlobalConfigurationDialogComponent } from "../../../globalSettings/components/globalConfigurationDialog/globalConfigurationDialog.component"
 import { SettingsButtonComponent } from "./settingsButton.component"
