+++ visualization/app/codeCharta/features/navBar/services/uploadFiles.service.ts
@@ -1,5 +1,5 @@
 import { Injectable } from "@angular/core"
-import { LoadFileService } from "../../../services/loadFile/loadFile.service"
+import { LoadFileService } from "../../../features/loadFile/facade"
 import { getCCFileAndDecorateFileChecksum } from "../../../util/fileHelper"
 import { createCCFileInput } from "../../../util/uploadFiles/createCCFileInput"
 import { readFiles } from "../../../util/uploadFiles/readFiles"
