--- visualization/app/codeCharta/services/loadFile/loadFile.service.ts
@@ -1,75 +1,0 @@
-import { Injectable, OnDestroy } from "@angular/core"
-import { tap } from "rxjs"
-import { clone } from "../../util/clone"
-import { CCFileValidationResult } from "../../util/fileValidator"
-import { setFiles, setStandardByNames } from "../../state/store/files/files.actions"
-import { FileState } from "../../model/files/files"
-import { NameDataPair, CcState } from "../../codeCharta.model"
-import { referenceFileSelector } from "../../state/selectors/referenceFile/referenceFile.selector"
-import { ErrorDialogService } from "../../features/shared/components/errorDialog/errorDialog.service"
-import { loadFilesValidationToErrorDialog } from "../../util/loadFilesValidationToErrorDialog"
-import { enrichFileStatesAndRecentFilesWithValidationResults } from "../../util/fileParser"
-import { fileRoot } from "../../util/fileRoot"
-import { Store, State } from "@ngrx/store"
-import { setCurrentFilesAreSampleFiles } from "../../state/store/appStatus/currentFilesAreSampleFiles/currentFilesAreSampleFiles.actions"
-
-export const NO_FILES_LOADED_ERROR_MESSAGE = "File(s) could not be loaded"
-export const FILES_ALREADY_LOADED_ERROR_MESSAGE = "File(s) are already loaded"
-
-@Injectable({ providedIn: "root" })
-export class LoadFileService implements OnDestroy {
-    referenceFileSubscription = this.store
-        .select(referenceFileSelector)
-        .pipe(
-            tap(newReferenceFile => {
-                if (newReferenceFile) {
-                    fileRoot.updateRoot(newReferenceFile.map.name)
-                }
-            })
-        )
-        .subscribe()
-
-    constructor(
-        private readonly store: Store<CcState>,
-        private readonly state: State<CcState>,
-        private readonly errorDialogService: ErrorDialogService
-    ) {}
-
-    ngOnDestroy(): void {
-        this.referenceFileSubscription.unsubscribe()
-    }
-
-    loadFiles(nameDataPairs: NameDataPair[]) {
-        const fileStates: FileState[] = clone(this.state.getValue().files)
-        const recentFiles: string[] = []
-        const fileValidationResults: CCFileValidationResult[] = []
-
-        const hasAddedAtLeastOneFile = enrichFileStatesAndRecentFilesWithValidationResults(
-            fileStates,
-            recentFiles,
-            nameDataPairs,
-            fileValidationResults,
-            () => this.state.getValue().appStatus.currentFilesAreSampleFiles,
-            () => this.store.dispatch(setCurrentFilesAreSampleFiles({ value: false }))
-        )
-
-        if (fileValidationResults.length > 0) {
-            this.errorDialogService.open(loadFilesValidationToErrorDialog(fileValidationResults))
-        }
-
-        if (recentFiles.length === 0) {
-            throw new Error(NO_FILES_LOADED_ERROR_MESSAGE)
-        }
-
-        this.store.dispatch(setFiles({ value: fileStates }))
-
-        const recentFile = recentFiles[0]
-        const rootName = this.state.getValue().files.find(f => f.file.fileMeta.fileName === recentFile).file.map.name
-        this.store.dispatch(setStandardByNames({ fileNames: recentFiles }))
-        fileRoot.updateRoot(rootName)
-
-        if (!hasAddedAtLeastOneFile) {
-            throw new Error(FILES_ALREADY_LOADED_ERROR_MESSAGE)
-        }
-    }
-}
