+++ visualization/app/codeCharta/state/effects/updateQueryParameters/updateQueryParameters.effect.ts
@@ -3,7 +3,7 @@ import { Actions, createEffect, ofType } from "@ngrx/effects"
 import { State, Store } from "@ngrx/store"
 import { debounceTime, map, tap, withLatestFrom } from "rxjs"
 import { CcState } from "../../../codeCharta.model"
-import { LoadInitialFileService } from "../../../services/loadInitialFile/loadInitialFile.service"
+import { LoadInitialFileService } from "../../../features/loadFile/facade"
 import { metricDataSelector } from "../../selectors/accumulatedData/metricData/metricData.selector"
 import { actionsRequiringUpdateQueryParameters } from "./actionsRequiringUpdateQueryParameters"
 import { MetricQueryParemter } from "./metricQueryParameter"
