+++ visualization/app/codeCharta/state/effects/updateQueryParameters/updateQueryParameters.spec.ts
@@ -5,7 +5,7 @@ import { Action, State } from "@ngrx/store"
 import { MockStore, provideMockStore } from "@ngrx/store/testing"
 import { waitFor } from "@testing-library/angular"
 import { Subject } from "rxjs"
-import { LoadInitialFileService } from "../../../services/loadInitialFile/loadInitialFile.service"
+import { LoadInitialFileService } from "../../../features/loadFile/facade"
 import { EDGE_METRIC_DATA, METRIC_DATA } from "../../../util/dataMocks"
 import { metricDataSelector } from "../../selectors/accumulatedData/metricData/metricData.selector"
 import { setColorMetric } from "../../store/dynamicSettings/colorMetric/colorMetric.actions"
